requirements.txt
Flask==3.1.3
supabase==2.28.0
python-dotenv==1.2.2
pylint==4.0.5
pytest==9.0.2
coverage==7.13.4
